backend/requirements.txt
# Backend dependencies for Blog Platform

# Flask core
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1
PyRSS2Gen==1.1
Pillow==10.1.0

# Testing
pytest==7.4.3
pytest-flask==1.3.0
